+++ packages/li-p2/src/LayerAttribute/components/ResterScaleSelector/CustomMappingColors/CustomInput/index.tsx
@@ -3,28 +3,25 @@ import { usePrefixCls } from '@formily/antd-v5/esm/__builtins__';
 import { ColorPicker } from 'antd';
 import type { Color } from 'antd/es/color-picker';
 import classnames from 'classnames';
-import React, { useMemo } from 'react';
-import Item from './Item';
+import React from 'react';
+import InputNumber from '../../../ScaleSelector/CustomMappingColors/CustomInput/InputNumber';
+import InputCat from './InputCat';
 import useStyle from './style';
 
 type RangeItemProps = {
-  dataType: 'string' | 'number';
+  customType: 'cat' | 'custom';
   position: string | null;
-  selectedOption: (string | number)[];
-  selectOptions: { label: string; value: string }[];
   min?: number;
   max?: number;
   color: string;
-  value: (string | number)[];
+  value: number[];
   onDelete: () => void;
-  onChange: (value: (string | number)[], color: string) => void;
+  onChange: (value: number[], color: string) => void;
 };
 
 const RangeItem = ({
-  dataType,
+  customType,
   position,
-  selectedOption,
-  selectOptions,
   color: defaultColor,
   value: defaultValue,
   min = 0,
@@ -32,23 +29,14 @@ const RangeItem = ({
   onDelete,
   onChange,
 }: RangeItemProps) => {
-  const prefixCls = usePrefixCls('formily-scale-selector__custom-content__custom-item');
+  const prefixCls = usePrefixCls('formily-rester--scale-selector__custom-input');
   const [wrapSSR, hashId] = useStyle(prefixCls);
-
-  const options = useMemo(() => {
-    if (!selectedOption.length && !defaultValue.length) {
-      return selectOptions;
-    }
-
-    const selected = selectedOption.filter((item) => item && !defaultValue.includes(item));
-    return selectOptions.filter((item) => !selected.includes(item.value));
-  }, [selectedOption, selectOptions, defaultValue]);
 
   const colorChange = (color: Color) => {
     onChange?.(defaultValue, color.toHexString());
   };
 
-  const onValueChange = (_value: (string | number)[]) => {
+  const onValueChange = (_value: number[]) => {
     onChange?.(_value, defaultColor);
   };
 
@@ -67,16 +55,20 @@ const RangeItem = ({
         </div>
 
         <div className={`${prefixCls}__infor__content`}>
-          <Item
-            size="small"
-            customType={dataType}
-            value={defaultValue}
-            options={options}
-            onChange={onValueChange}
-            min={min}
-            max={max}
-            position={position}
-          />
+          {customType === 'cat' && (
+            <InputCat size="small" min={min} max={max} value={defaultValue as [number]} onChange={onValueChange} />
+          )}
+
+          {customType === 'custom' && (
+            <InputNumber
+              size="small"
+              value={defaultValue as [number, number]}
+              min={min}
+              max={max}
+              position={position}
+              onChange={onValueChange}
+            />
+          )}
         </div>
 
         <div className={`${prefixCls}__infor__delete-icon`} onClick={onDelete}>
