+++ packages/li-p2/src/LayerAttribute/components/ScaleSelector/CustomMappingColors/CustomInput/InputNumber/style.ts
@@ -1,6 +1,6 @@
 import { genStyleHook } from '@formily/antd-v5/esm/__builtins__';
 
-export default genStyleHook('scale-selector__custom-content__custom-item__item-number', (token) => {
+export default genStyleHook('scale-selector__custom-input__number', (token) => {
   const { componentCls } = token;
 
   return {
