+++ packages/li-p2/src/LayerAttribute/components/ScaleSelector/CustomMappingColors/CustomInput/InputString/index.tsx
